Even smarter deploy

+++ deploy.sh
@@ -1,17 +1,23 @@
 #!/bin/bash
 
-echo Copying peerflix server
-cp ~/.npm-packages/lib/node_modules/peerflix-server/server/index.js peerflix-index.js
-
-echo Running build
+echo "--- Running build ---"
 npm run build
 
-echo Removing old version
+echo "--- Removing old version ---"
 ssh -t up "rm -rf ~/yify-react/*"
 
-echo Copying new build/server
+echo "--- Copying new build ---"
 scp -r build up:~/yify-react/
-scp peerflix-index.js up:~/index.js
 
-echo Updating peerflix container
-ssh -t up "docker cp ~/index.js peerflix:/home/app/server/ && rm ~/index.js && docker restart peerflix"
+# Check if the peerflix server has changed at all
+A=~/.npm-packages/lib/node_modules/peerflix-server/server/index.js
+B=peerflix-index.js
+DIFF=$(diff $A $B)
+if [ "$DIFF" ]; then
+    echo "--- Copying peerflix server ---"
+    cp $A $B
+    scp peerflix-index.js up:~/index.js
+
+    echo "--- Updating peerflix container ---"
+    ssh -t up "docker cp ~/index.js peerflix:/home/app/server/ && rm ~/index.js && docker restart peerflix"
+fi
